Moved Request class to new sub-package data created GameStatusList data class
