+++ app/src/main/java/dk/kaddu/phoenixbsecompanion/data/GameStatusList.kt
@@ -1,0 +1,7 @@
+package dk.kaddu.phoenixbsecompanion.data
+
+data class GameStatusList (val gameStatus: List<GameStatus>)
+
+data class GameStatus (
+        val star_date: String?,
+        val status: String?)
